many improvements
I made changes in the RPCBuilders.go file (it did not compile):
I kept the idea and the structure
used go conventions in line breaks, capital letters,...
fixed some mistakes
change the name of the file to rpcBuilders.go, without capital letter, based on convention
I added a layer (walletdrpc.go) for easy handling of the communication with walletd and easy integration in an app
I changed the package name. We cannot have capital letters or dash in a go package name. A,d it is better if the name of the repo is the name of the package (for easy install with go get github.com/turtlecoin/turtlecoinwalletdrpcgo). I don't know if the name of the repo can still be changed?
I use this version in my app, so I confirm it works as expected